Onchain-data/Afterglow36/TronWhaleAgent/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.51.0
requests>=2.31.0,<3.0.0